colorful_map/backend/requirements.txt
django>=6.0.2
channels
daphne
djangorestframework
django-cors-headers
